extract Subheading component
