+++ dotcom-rendering/src/components/KeyTakeaway.tsx
@@ -1,12 +1,11 @@
 import { css } from '@emotion/react';
 import { type ArticleFormat } from '@guardian/libs';
-import { remSpace } from '@guardian/source-foundations';
 import type { EditionId } from '../lib/edition';
 import type { ArticleElementRenderer } from '../lib/renderElement';
 import { palette } from '../palette';
 import type { ServerSideTests, Switches } from '../types/config';
 import type { KeyTakeaway as KeyTakeawayModel } from '../types/content';
-import { subheadingStyles } from './SubheadingBlockComponent';
+import { Subheading } from './Subheading';
 
 const keyTakeawayStyles = css`
 	padding-top: 8px;
@@ -60,15 +59,10 @@ export const KeyTakeaway = ({
 		<>
 			<li css={keyTakeawayStyles}>
 				<hr css={headingLineStyles} />
-				<h2
-					css={[
-						subheadingStyles(format),
-						`padding: ${remSpace[0]} 0;`,
-					]}
-				>
+				<Subheading format={format} topPadding={false}>
 					<span css={headingIndexStyles}>{`${titleIndex}. `}</span>
 					{keyTakeaway.title}
-				</h2>
+				</Subheading>
 				{keyTakeaway.body.map((element, index) => (
 					<RenderArticleElement
 						// eslint-disable-next-line react/no-array-index-key -- This is only rendered once so we can safely use index to suppress the warning
